@@ -3,12 +3,14 @@
 namespace Agenciafmd\Postal\Notifications;
 
 use Agenciafmd\Leads\Channels\LeadChannel;
-use Agenciafmd\Postal\Mails\SendMail;
+use Agenciafmd\Leads\Models\Lead;
+use Agenciafmd\Postal\Models\Postal;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Channels\MailChannel;
+use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
-use Agenciafmd\Leads\Models\Lead;
+use Symfony\Component\Mime\Email;
 
 class SendNotification extends Notification implements ShouldQueue
 {
@@ -16,46 +18,101 @@ class SendNotification extends Notification implements ShouldQueue
 
     public int $tries = 3;
 
-    public $data;
+    public function __construct(
+        public array $data = [],
+        public array $from = [],
+        public array $attach = [],
+        public ?string $subject = null,
+    ) {}
 
-    public $from;
-
-    public $attach;
-
-    public $subject;
-
-    public function __construct($data, $from = null, $attach = null, $subject = null)
+    /* TODO: verificar se conseguimos criar um evento customizado a fim de disparar o leadChannel lá no pacote de leads */
+    public function via(Postal $notifiable): array
     {
-        $this->data = $data;
+        $leadChannel = class_exists(LeadChannel::class) ? [
+            LeadChannel::class,
+        ] : [];
 
-        $this->from = $from;
-
-        $this->attach = $attach;
-
-        $this->subject = $subject;
-    }
-
-    public function via($notifiable)
-    {
         return [
             MailChannel::class,
-            LeadChannel::class,
+            ...$leadChannel,
         ];
     }
 
-    public function toMail($notifiable)
+    public function toMail(Postal $notifiable): MailMessage
     {
-        return new SendMail($notifiable, $this->from, $this->data, $this->attach, $this->subject);
+        $content = array_merge([
+            'greeting' => __('Hi :name!', ['name' => $notifiable->to_name]),
+            'introLines' => [
+                __('This email sent by the website through the :name form.', ['name' => $notifiable->name]),
+            ],
+            'outroLines' => [
+            ],
+        ], $this->data);
+
+        $mail = (new MailMessage)
+            ->markdown('admix-mail::markdown.email')
+            ->theme('admix-mail::theme.tabler')
+            ->level('default')
+            ->subject(($this->subject) ?? config('app.name') . ' | ' . $notifiable->subject);
+
+        if ($content['greeting']) {
+            $mail->greeting($content['greeting']);
+        }
+
+        foreach ($content['introLines'] as $introLine) {
+            $mail->line($introLine);
+        }
+
+        if ($content['actionText'] && $content['actionUrl']) {
+            $mail->action($content['actionText'], $content['actionUrl']);
+        }
+
+        foreach ($content['outroLines'] as $outroLine) {
+            $mail->line($outroLine);
+        }
+
+        if ($this->from) {
+            $mail->replyTo(key($this->from), current($this->from));
+        }
+
+        if ($cc = $notifiable->cc) {
+            $ccs = array_map('trim', explode(',', $cc));
+            foreach ($ccs as $cc) {
+                $mail->cc($cc);
+            }
+        }
+
+        if ($bcc = $notifiable->bcc) {
+            $bccs = array_map('trim', explode(',', $bcc));
+            foreach ($bccs as $bcc) {
+                $mail->bcc($bcc);
+            }
+        }
+
+        /* TODO: modificar para o attachFromStorage quando subir a versão do laravel */
+        foreach ($this->attach as $attach) {
+            $mail->attach($attach);
+        }
+
+        $mail->withSymfonyMessage(static function (Email $message) {
+            $message->getHeaders()
+                ->addTextHeader(
+                    'X-Mailgun-Tag', config('app.name')
+                );
+        });
+
+        return $mail;
     }
 
-    public function toLead($data)
+    public function toLead(array $data): void
     {
-        Lead::create([
-            'source' => $data['source'],
-            'name' => $data['name'],
-            'email' => $data['email'],
-            'phone' => $data['phone'],
-            'description' => $data['message'],
-        ]);
+        Lead::query()
+            ->create([
+                'source' => $data['source'],
+                'name' => $data['name'],
+                'email' => $data['email'],
+                'phone' => $data['phone'],
+                'description' => $data['message'],
+            ]);
     }
 }
